students-support-system-jaist/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
pytesseract==0.3.10
Pillow==10.2.0
